requirements.txt
# LPCA Requirements
# Core dependencies for running experiments

# Deep Learning
torch>=2.2.0
transformers>=4.38.0
accelerate>=0.27.0
datasets>=2.18.0
safetensors>=0.4.0
einops>=0.7.0

# Scientific Computing
numpy>=1.26.0
scipy>=1.12.0
pandas>=2.2.0
pyarrow>=15.0.0

# Configuration & Utilities
pydantic>=2.6.0
pyyaml>=6.0.0
tqdm>=4.66.0
rich>=13.7.0

# Visualization (optional but recommended)
matplotlib>=3.8.0
seaborn>=0.13.0

# Machine Learning Utilities
scikit-learn>=1.4.0

# Development Tools (optional)
pytest>=8.0.0
black>=24.0.0
ruff>=0.2.0

# Experiment Tracking (optional)
# wandb>=0.16.0
# hydra-core>=1.3.0
